+++ Algorithm.CSharp/BasicTemplateAlgorithm.cs
