+++ Common/Interfaces/IApi.cs
@@ -263,13 +263,6 @@ namespace QuantConnect.Interfaces
         void SetAlgorithmStatus(string algorithmId, AlgorithmStatus status, string message = "");
 
         /// <summary>
-        /// Will get the prices for requested symbols
-        /// </summary>
-        /// <param name="symbols">Symbols for which the price is requested</param>
-        /// <returns><see cref="Prices"/></returns>
-        PricesList ReadPrices(IEnumerable<Symbol> symbols);
-
-        /// <summary>
         /// Send the statistics to storage for performance tracking.
         /// </summary>
         /// <param name="algorithmId">Identifier for algorithm</param>
@@ -291,22 +284,6 @@ namespace QuantConnect.Interfaces
         /// <param name="subject">The email subject</param>
         /// <param name="body">The email message body</param>
         void SendUserEmail(string algorithmId, string subject, string body);
-
-        /// <summary>
-        /// Gets all split events between the specified times. From and to are inclusive.
-        /// </summary>
-        /// <param name="from">The first date to get splits for</param>
-        /// <param name="to">The last date to get splits for</param>
-        /// <returns>A list of all splits in the specified range</returns>
-        List<Data.Market.Split> GetSplits(DateTime from, DateTime to);
-
-        /// <summary>
-        /// Gets all dividend events between the specified times. From and to are inclusive.
-        /// </summary>
-        /// <param name="from">The first date to get dividend for</param>
-        /// <param name="to">The last date to get dividend for</param>
-        /// <returns>A list of all dividend in the specified range</returns>
-        List<Data.Market.Dividend> GetDividends(DateTime from, DateTime to);
 
         /// <summary>
         /// Local implementation for downloading data to algorithms
